Refactor FunctionCallInsn class
